Move cron aspects into its own file, add provisions for multi-threaded startup
